+++ src/features/Cameras/scripts/lib/initDetector.ts
