+++ src/features/Cameras/scripts/utils/elementTypeCheck.ts
